+++ tests/test_cancel_scopes.py
@@ -9,17 +9,30 @@ async def test_move_on_when(autojump_clock):
     event = trio.Event()
 
     async with move_on_when(event.wait) as cancel_scope:
-        assert isinstance(cancel_scope, trio.CancelScope)
+        pass
         # context manager exits normally even if awaitable doesn't return
+    assert not cancel_scope.cancel_called
+    assert not cancel_scope.cancelled_caught
 
     set_event = False
-    async with move_on_when(event.wait):
+    async with move_on_when(event.wait) as cancel_scope:
         await trio.sleep(1)
         assert not event.is_set()
         event.set()
         set_event = True
         await trio.sleep_forever()
     assert set_event
+    assert cancel_scope.cancel_called
+    assert cancel_scope.cancelled_caught
+
+
+async def test_move_on_when_deadline(autojump_clock):
+    async with move_on_when(trio.sleep_forever) as cancel_scope:
+        cancel_scope.deadline = trio.current_time() + 1
+        await trio.sleep_forever()
+    assert trio.current_time() == 1
+    assert cancel_scope.cancel_called
+    assert cancel_scope.cancelled_caught
 
 
 async def test_move_on_when_args():